Convert exceptions to ConflictException (#213)
A conflict can occur if multiple users arre trying to change the state of the same Dialogmote at the same tid. In case of such a conflict, throw a ConflictException instead of a generic RuntimeException, and return a http 409 indicating a conflict in the state of the accessed Dialogmote. Then the consuming application can update its state accordingly.