Turn of Dialogmoteinnkalling for persons with inactive Oppfolgingstilfelle
If a person has an Oppfolgingstilfelle that hasn't been active for the last 4 month, then the Narmeste Leder of the person will not have access to Dine Sykmeldte and hence no access to view Dialogmote-brev. A check raising an exception will prevent a Veileder from creating a Dialogmoteinnkalling where the latest Oppfolgingstilfelle of a person is inactive. This will prevent Dialogmoter where Narmeste Leder(Arbeidsgiver) without access to Dialogmoter.