Do not invalidate a late behandler svar (#417)
Remove concept that a behandler svar can be invalid and put all behandler
svar on kafka. This means that a late behandler svar, i.e we
recieve it after a meeting is closed, must be handled by the consumer (if
it deems it necessary).

The only consumer currently is ispersonoppgave, which means this commit
must be merged after it's handled in ispersonoppgave.