IS-1867: Add eksternReferanseId to dokarkiv body (#480)
This will be required by dokarkiv 20. december.
If api receives an eksternReferanseId that is already stored, it will
return 409 Conflict with the existing JournalPost as body.
We want to fail when we get 409 becuase we don't want to journalføre the
same varsel more than once. Since 409 gives clientException, the
existing client returns null, which is then handled by the cron jobs.
In referat create UUID for eksternReferanseId based on uuid and
recipient, to make sure it's unique for each journalfort version, and
also the same each time the same document is journalfort.